# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Todo List Api Version 1",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/sanctum/token": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Create JWT token",
                "operationId": "e0a0a36369bd835ea2890ffe60a9ffb1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "test1@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "string",
                                                            "example": "5fa8hgd2-563c-4rtf-bb0a-ab51967f456c"
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "John Doe"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "test1@example.com"
                                                        },
                                                        "email_verified_at": {
                                                            "type": "string",
                                                            "example": "2023-11-20T12:00:00Z"
                                                        },
                                                        "created_at": {
                                                            "type": "string",
                                                            "example": "2023-11-20T12:00:00Z"
                                                        },
                                                        "updated_at": {
                                                            "type": "string",
                                                            "example": "2023-11-20T12:00:00Z"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "authorization": {
                                                    "properties": {
                                                        "token": {
                                                            "type": "string",
                                                            "example": "your_jwt_token_here"
                                                        },
                                                        "type": {
                                                            "type": "string",
                                                            "example": "bearer"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/422"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/401"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/sanctum/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout user",
                "operationId": "5e6c69da3b03e68cad8324716a278e35",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "success": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/401"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "401": {
                "title": "Unauthenticated access",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
...
            "422": {
                "title": "Validation error response",
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "errors": {
                        "properties": {
                            "email": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            },
                            "password": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "token": {
                "type": "http",
                "description": "Enter Bearer Token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}